bootc-playground
I am going through a lot of technologies and documentation that are
largely new to me; at least using them is new.  To make my life easier
and automate as many things away, I started writing a Makefile that may
be useful for others interested in playing with and learning about
bootc.  I added comments and links to docs to leave breadcrumbs for
others in case they need or want to extend the workflow.

I plan on extending the docs and the Makefile once I gained more
experience using bootc.

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>